backend/requirements.txt
fastapi==0.110.0
uvicorn==0.30.1
pymongo[srv]==4.7.2
python-dotenv==1.0.1
pydantic==2.7.1
